# Simplify building Hooks and ContentTypeParser