--- a/lib/ContentTypeParser.js
+++ b/lib/ContentTypeParser.js
@@ -70,12 +70,10 @@ ContentTypeParser.prototype.run = function (contentType, handler, request, reply
 }
 
 function buildContentTypeParser (c) {
-  function _ContentTypeParser () {}
-  _ContentTypeParser.prototype = new ContentTypeParser()
-  const C = new _ContentTypeParser()
-  C.customParsers = Object.create(c.customParsers)
-  C.parserList = Object.create(c.parserList)
-  return C
+  const contentTypeParser = new ContentTypeParser()
+  Object.assign(contentTypeParser.customParsers, c.customParsers)
+  contentTypeParser.parserList = c.parserList.slice()
+  return contentTypeParser
 }
 
 module.exports = ContentTypeParser