--- a/lib/hooks.js
+++ b/lib/hooks.js
@@ -26,14 +26,12 @@ Hooks.prototype.add = function (hook, fn) {
 }
 
 function buildHooks (h) {
-  function _Hooks () {}
-  _Hooks.prototype = new Hooks()
-  const H = new _Hooks()
-  H.onRequest = Object.create(h.onRequest)
-  H.preHandler = Object.create(h.preHandler)
-  H.onResponse = Object.create(h.onResponse)
-  H.onSend = Object.create(h.onSend)
-  return H
+  const hooks = new Hooks()
+  hooks.onRequest = h.onRequest.slice()
+  hooks.preHandler = h.preHandler.slice()
+  hooks.onSend = h.onSend.slice()
+  hooks.onResponse = h.onResponse.slice()
+  return hooks
 }
 
 module.exports = Hooks