Creating a new constructor and inheriting from the old one isn't necessary for `Hooks` and `ContentTypeParser`. Simply creating a new instance and cloning the values is all that's needed.

This also improves debugging, general code compatibility, and possibly performance since the hooks and `parsersList` will now be real arrays instead of array-like objects.