@nwoltman 2018-01-03
Benchmarks aren't very interesting:

**master**

```
[1] Stat         Avg     Stdev   Max     
[1] Latency (ms) 3.57    10.51   206     
[1] Req/Sec      27265.6 4654.58 30111   
[1] Bytes/Sec    4.1 MB  717 kB  4.72 MB 
[1] 
[1] 136k requests in 5s, 20.3 MB read
...
[1] Stat         Avg     Stdev   Max     
[1] Latency (ms) 3.43    10.01   197     
[1] Req/Sec      28395.2 5027.06 32159   
[1] Bytes/Sec    4.23 MB 727 kB  4.98 MB 
[1] 
[1] 142k requests in 5s, 21.2 MB read
```

**this PR**

```
[1] Stat         Avg     Stdev   Max     
[1] Latency (ms) 3.58    10.26   222     
[1] Req/Sec      27204.8 4396.45 30047   
[1] Bytes/Sec    4.09 MB 659 kB  4.72 MB 
[1] 
[1] 136k requests in 5s, 20.3 MB read
...
[1] Stat         Avg     Stdev   Max     
[1] Latency (ms) 3.38    10      210     
[1] Req/Sec      28755.2 5845.06 33183   
[1] Bytes/Sec    4.29 MB 873 kB  4.98 MB 
[1] 
[1] 144k requests in 5s, 21.4 MB read
```